fixes #1

+++ test-host/test_app_mqtt.cc
@@ -6,12 +6,11 @@
 using HippoMocks::CString;
 
 extern "C" {
-  bool getTemperatureValue(unsigned int* value, const cJSON* root, const char* tag);
+  bool getTemperatureValue(short* value, const cJSON* root, const char* tag);
 }
 
 TEST_CASE("test get ", "[tag]" ) {
-
-  unsigned int temp=0;
+  short temp=0;
   const char* json = "{\"humidity\":53.3,\"temperature\":18.3}";
   cJSON * root   = cJSON_Parse(json);
